% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPN-package.R
\docType{package}
\name{CPN-package}
\alias{CPN}
\alias{CPN-package}
\title{CPN: Compound Poisson-Normal Regression}
\description{
Provides tools to fit regression models when the outcome follows a Compound Poisson-Normal (CPN) distribution. The model assumes a Poisson-distributed count of latent normal variables whose sum is observed, allowing joint modeling of zero inflation and continuous positive values. Includes methods for estimation, inference, simulation, and model diagnostics.
}
\author{

\strong{Maintainer}: Laszlo Pecze \email{laszlopecze77@gmail.com} (\href{https://orcid.org/0000-0002-7036-5756}{ORCID})

Other contributors:
\itemize{
  \item François Collin \email{fca.collin@gmail.com} (\href{https://orcid.org/0000-0003-0524-5755}{ORCID}) [contributor]
}


}
\keyword{internal}
