% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpn.R
\name{cpn}
\alias{cpn}
\title{Compound Poisson-Normal Regression}
\usage{
cpn(formula, data = NULL, mu_init = NULL, sigma_init = NULL, k_max = 10)
}
\arguments{
\item{formula}{A formula specifying the model, e.g., \code{y ~ x1 + x2}.}

\item{data}{An optional data frame, list, or environment containing the
variables in the model. If not found in \code{data}, the variables are
taken from the environment from which \code{cpn} is called.}

\item{mu_init}{Optional initial value for the Normal mean parameter.
If \code{NULL}, it is set to the sample mean of \code{y}.}

\item{sigma_init}{Optional initial value for the Normal standard deviation.
If \code{NULL}, it is set to the sample standard deviation of \code{y}.}

\item{k_max}{Upper limit of summation used in approximating the Poisson
convolution. Should be between 10 and 100. Default is 10.}
}
\value{
An object of class \code{"cpn"} containing the following components:
\item{coefficients}{Estimated regression coefficients.}
\item{mu}{Estimated mean of the Normal component.}
\item{sigma}{Estimated standard deviation of the Normal component.}
\item{se}{Standard errors of the estimated parameters.}
\item{model}{The model frame used.}
\item{terms}{The terms object from the model.}
\item{formula}{The model formula.}
\item{data}{The original data passed in.}
\item{deviance_residuals}{Vector of deviance residuals.}
\item{fitted_values}{Fitted values (Poisson mean times Normal mean).}
\item{neg_log_likelihood}{Negative log-likelihood at the optimum.}
\item{null_deviance}{Null model deviance.}
\item{residual_deviance}{Residual deviance of the fitted model.}
\item{df_null}{Degrees of freedom for the null model.}
\item{df_residual}{Degrees of freedom for the fitted model.}
\item{k_max}{Value used to truncate the Poisson convolution sum.}
\item{call}{The matched function call.}
}
\description{
Fits a Compound Poisson-Normal (CPN) regression model to the response
variable using maximum likelihood estimation via the Nelder-Mead method
\insertCite{Nelder1965}{CPN}.
}
\details{
The function fits a regression model where the response is assumed to follow
a Compound Poisson-Normal distribution. The model estimates regression
coefficients for the Poisson intensity, and mean and standard deviation of
the Normal component.

The optimization is performed via maximum likelihood using the Nelder-Mead
method. Standard errors are computed via the observed information matrix
using numerical Hessian. If the Hessian is singular or not positive
definite, a warning is issued and standard errors are returned as \code{NA}.
}
\examples{
set.seed(123)
data <- simulate_cpn_data()

# Sequential analysis of deviance
fit <- cpn(y ~ x1 + x2, data = data)
summary(fit)

}
\references{
\insertRef{Nelder1965}{CPN}
}
