% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.cpn}
\alias{logLik.cpn}
\title{Extract Log-Likelihood from a CPN Model}
\usage{
\method{logLik}{cpn}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"cpn"}, typically resulting from a
CPN model fitting function.}

\item{...}{Additional arguments (currently unused).}
}
\value{
An object of class \code{"logLik"} representing the log-likelihood.
The \code{df} attribute contains the number of estimated parameters.
}
\description{
Returns the log-likelihood of a fitted Compound Poisson-Normal (CPN) model.
}
\seealso{
\code{\link{AIC}}, \code{\link{BIC}}, \code{\link{logLik}}
}
