% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_data.R
\name{recover_data.cpn}
\alias{recover_data.cpn}
\title{Recover Data for \code{cpn} Model Objects}
\usage{
recover_data.cpn(object, ...)
}
\arguments{
\item{object}{A fitted model object of class \code{cpn}.}

\item{...}{Additional arguments (not used).}
}
\value{
An object containing the recovered data and terms, suitable for use
by \pkg{emmeans}.
}
\description{
This method is used by the \pkg{emmeans} package to recover the data and
terms from a fitted Compound Poisson-Normal (CPN) model object. It is
typically used internally when computing estimated marginal means.
}
