% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.cpn}
\alias{anova.cpn}
\title{Analysis of Deviance for CPN Models}
\usage{
\method{anova}{cpn}(object, ..., test = "Chisq")
}
\arguments{
\item{object}{An object of class \code{cpn}, typically the result of a call to \code{\link[=cpn]{cpn()}}.}

\item{...}{Optional additional objects of class \code{cpn} for model comparison.}

\item{test}{A character string specifying the test statistic. Currently only \code{"Chisq"} is supported.}
}
\value{
An object of class \code{"anova"} inheriting from \code{"data.frame"}, containing:
\itemize{
\item If one model is provided: sequential deviance table by added terms.
\item If multiple models are provided: deviance comparison between models.
}
}
\description{
Computes an analysis of deviance table for objects of class \code{cpn}, either:
(1) sequentially by model terms (Type I ANOVA) when one model is supplied, or
(2) by comparing two or more nested models.
}
\details{
When a single model is supplied, this function refits the model with terms
added sequentially and reports the change in residual deviance.

When multiple models are supplied, they are assumed to be nested and ordered
by increasing complexity (based on residual degrees of freedom). The function
then compares their deviances using chi-squared tests.
}
\examples{
\dontrun{
# Sequential analysis of deviance
fit <- cpn(y ~ x1 + x2, data = dat)
anova(fit)

# Model comparison
fit1 <- cpn(y ~ x1, data = dat)
fit2 <- cpn(y ~ x1 + x2, data = dat)
anova(fit1, fit2)
}

}
